backend/requirements.txt
# Python 3.11.x Required for optimal LLM performance and compatibility
# Ensure you're using Python 3.11.x before installing dependencies

# Core Framework
fastapi==0.110.3
uvicorn[standard]==0.27.1
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# AI/ML
click==8.1.7
crewai==0.130.0 
crewai-tools==0.47.1
google-ai-generativelanguage==0.6.4
google-api-core==2.10.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.53.0
google-cloud-bigquery==3.23.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.5.4
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
Jinja2==3.1.4
jsonschema==4.22.0
langchain-core==0.1.52
langsmith==0.1.67
numpy>=1.26.0
oauthlib==3.2.2
onnxruntime>=1.20.0
openai==1.30.5
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
pandas==2.2.2
pillow==10.3.0
pip==24.0
protobuf==4.25.3
pydantic==1.10.13
pydantic_core==2.8.0
# File processing dependencies
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-multipart==0.0.6
uvicorn==0.27.1
# Database
pymongo==4.6.1
motor==3.3.2
dnspython==2.4.2
# Authentication dependencies
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator==2.1.0

# Redis caching and background job processing
redis==5.0.1
celery==5.3.4
flower==2.0.1

# Rate limiting
slowapi==0.1.9

# LLM Observability and Monitoring
wandb==0.16.1
tiktoken==0.5.2

# Export functionality dependencies
reportlab==4.0.7

# Advanced file processing dependencies
pdfplumber==0.10.3
pypdf==4.0.1
pytesseract==0.3.10
PyMuPDF==1.23.14
python-magic-bin==0.4.14

# Security and 2FA dependencies
pyotp==2.9.0
qrcode==7.4.2
bcrypt==4.1.2